Removed "Tests passed" section, useless and possibly confusing.
Links to "Kambi VRML test suite" and "NIST conformance test" are clearly visible on the sidebar.

"Chaco test suite" is rather unknown, and it's actually only a few models...

The fact that we handle Blender exported models is obvious. We handle models exported by any exporter of any 3D modeller...

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/www@7297 afe8dcc5-9094-451b-8773-5f495042a737